[syft] move service/response.py to types
