+++ packages/syft/src/syft/external/oblv/oblv_keys_stash.py
@@ -9,13 +9,13 @@ from result import Result
 
 # relative
 from ...serde.serializable import serializable
-from ...service.response import SyftError
 from ...store.document_store import BaseStash
 from ...store.document_store import DocumentStore
 from ...store.document_store import PartitionSettings
 from ...store.document_store import QueryKeys
 from ...store.document_store import UIDPartitionKey
 from ...types.credentials import SyftVerifyKey
+from ...types.response import SyftError
 from ...types.uid import UID
 from .oblv_keys import OblvKeys
