+++ packages/syft/src/syft/service/data_subject/data_subject_service.py
@@ -14,9 +14,9 @@ from ...store.document_store import PartitionSettings
 from ...store.document_store import QueryKeys
 from ...types.context import AuthedServiceContext
 from ...types.credentials import SyftVerifyKey
+from ...types.response import SyftError
+from ...types.response import SyftSuccess
 from ...util.telemetry import instrument
-from ..response import SyftError
-from ..response import SyftSuccess
 from ..service import AbstractService
 from ..service import SERVICE_TO_TYPES
 from ..service import TYPE_TO_SERVICE
