+++ packages/syft/src/syft/service/network/network_service.py
@@ -23,6 +23,8 @@ from ...store.document_store import QueryKeys
 from ...types.context import AuthedServiceContext
 from ...types.credentials import SyftVerifyKey
 from ...types.grid_url import GridURL
+from ...types.response import SyftError
+from ...types.response import SyftSuccess
 from ...types.transforms import TransformContext
 from ...types.transforms import keep
 from ...types.transforms import transform
@@ -31,8 +33,6 @@ from ...types.user_roles import GUEST_ROLE_LEVEL
 from ...util.telemetry import instrument
 from ..data_subject.data_subject import NamePartitionKey
 from ..metadata.node_metadata import NodeMetadata
-from ..response import SyftError
-from ..response import SyftSuccess
 from ..service import AbstractService
 from ..service import SERVICE_TO_TYPES
 from ..service import TYPE_TO_SERVICE
