+++ packages/syft/src/syft/service/policy/policy.py
@@ -33,6 +33,8 @@ from ...types.context import AuthedServiceContext
 from ...types.context import NodeServiceContext
 from ...types.credentials import SyftVerifyKey
 from ...types.datetime import DateTime
+from ...types.response import SyftError
+from ...types.response import SyftSuccess
 from ...types.syft_object import SYFT_OBJECT_VERSION_1
 from ...types.syft_object import SyftObject
 from ...types.transforms import TransformContext
@@ -45,8 +47,6 @@ from ..action.twin_object import TwinObject
 from ..code.code_parse import GlobalsVisitor
 from ..code.unparse import unparse
 from ..dataset.dataset import Asset
-from ..response import SyftError
-from ..response import SyftSuccess
 
 PolicyUserVerifyKeyPartitionKey = PartitionKey(
     key="user_verify_key", type_=SyftVerifyKey
