+++ packages/syft/src/syft/service/request/request_service.py
@@ -11,6 +11,8 @@ from ...serde.serializable import serializable
 from ...store.document_store import DocumentStore
 from ...store.linked_obj import LinkedObject
 from ...types.context import AuthedServiceContext
+from ...types.response import SyftError
+from ...types.response import SyftSuccess
 from ...types.uid import UID
 from ...types.user_roles import GUEST_ROLE_LEVEL
 from ...util.telemetry import instrument
@@ -19,8 +21,6 @@ from ..action.action_permissions import ActionPermission
 from ..message.message_service import CreateMessage
 from ..message.message_service import Message
 from ..message.message_service import MessageService
-from ..response import SyftError
-from ..response import SyftSuccess
 from ..service import AbstractService
 from ..service import SERVICE_TO_TYPES
 from ..service import TYPE_TO_SERVICE
