+++ packages/syft/tests/syft/api_test.py
@@ -8,8 +8,8 @@ import pytest
 
 # syft absolute
 import syft as sy
-from syft.service.response import SyftAttributeError
 from syft.service.user.user import UserUpdate
+from syft.types.response import SyftAttributeError
 from syft.types.user_roles import ServiceRole
