+++ packages/syft/tests/syft/messages/message_service_test.py
@@ -9,14 +9,14 @@ from syft.service.message.message_stash import MessageStash
 from syft.service.message.messages import CreateMessage
 from syft.service.message.messages import Message
 from syft.service.message.messages import MessageStatus
-from syft.service.response import SyftError
-from syft.service.response import SyftSuccess
 from syft.store.document_store import DocumentStore
 from syft.store.linked_obj import LinkedObject
 from syft.types.context import AuthedServiceContext
 from syft.types.credentials import SyftSigningKey
 from syft.types.credentials import SyftVerifyKey
 from syft.types.datetime import DateTime
+from syft.types.response import SyftError
+from syft.types.response import SyftSuccess
 from syft.types.uid import UID
 
 test_verify_key_string = (
